+++ app/src/main/java/com/madarsoft_task/features/common/domain/model/ui/UserUiModel.kt
@@ -3,7 +3,7 @@ package com.madarsoft_task.features.common.domain.model.ui
 import com.madarsoft_task.features.common.data.locale.entity.UserEntity
 
 data class UserUiModel(
-    val userId : Int? = null,
+    val userId : Long? = null,
     val name: String? = null,
     val age: Int? = null,
     val jobTitle: String? = null,
